refactor(remote-host): extract transport into @mulmoclaude/core
The remote-host command channel + host runner + connect lifecycle were a
per-host copy (itself ported from ../mulmoserver). MulmoTerminal is about to
become a second host, so extract the generic transport into @mulmoclaude/core
before a third copy exists.

Two new subpaths, split on the browser-safe boundary (mirrors ./remote-view):

- @mulmoclaude/core/remote-host (browser-safe): protocol wire types + Firestore
  path helpers, now parameterized by the firestore instance instead of a
  module-level singleton, so one module serves every host's own Firebase init.
- @mulmoclaude/core/remote-host/server: startHostRunner (command loop, heartbeat
  now an option), createRemoteHost (connect/disconnect/status lifecycle, hostId +
  logger injected so it imports no Firebase and stays unit-testable), plus
  createRemoteHostAuth / createRemoteHostFirebase init primitives.

firebase becomes an optional peerDependency of core (externalized in the build).
server/remoteHost/ is now a thin consumer: it binds MulmoClaude's firebaseConfig,
hostId="mulmoclaude", handler table, and logger to the shared engine. Behavior is
unchanged — the lifecycle test moved to core (generalized for injected hostId)
and all 88 host remote-host tests stay green.

Bumps @mulmoclaude/core 0.8.2 -> 0.9.0 (new subpaths) and the launcher +
collection-plugin peer core ranges to ^0.9.0 to keep the launcher-sync
invariant green. Publishing core@0.9.0 unblocks the MulmoTerminal port
(receptron/mulmoterminal#216).
