+++ packages/plugins/collection-plugin/package.json
@@ -31,13 +31,13 @@
     "zod": "^4.4.3"
   },
   "peerDependencies": {
-    "@mulmoclaude/core": "^0.8.0",
+    "@mulmoclaude/core": "^0.9.0",
     "gui-chat-protocol": "^0.4.0",
     "vue": "^3.5.0",
     "vue-i18n": "^11.4.4"
   },
   "devDependencies": {
-    "@mulmoclaude/core": "^0.8.0",
+    "@mulmoclaude/core": "^0.9.0",
     "@tailwindcss/vite": "^4.3.2",
     "@types/node": "^26.1.0",
     "@vitejs/plugin-vue": "^6.0.5",
